Tons of things
~57% performance increase -> from 4.5s to 1.9s for Chapter 7 in 7680 x 4320

Chapter 9 done
Chapter 10 stripes pattern done

+++ SourceHeaders/Utils.h
@@ -18,7 +18,7 @@ namespace Utils
 			T* newAlloc = new T[m_size + 1];
 
 			// moves old vector to new vector
-			for (size_t i = 0; i < m_size; i++)
+			for (uint32_t i = 0; i < m_size; i++)
 			{
 				newAlloc[i] = std::move(m_data[i]);
 			}
@@ -33,7 +33,7 @@ namespace Utils
 			T* newAlloc = new T[m_size + 1u];
 
 			// moves old vector to new vector
-			for (size_t i = 0; i < m_size; i++)
+			for (uint32_t i = 0; i < m_size; i++)
 			{
 				newAlloc[i] = std::move(m_data[i]);
 			}
